Laying some outline for file-based type detection
Since X-Plane plugins can be...messy...and to a computer they all look similar, I want to try including a JSON file (XPMod.json) in each plugin package that includes some info about the plugin.  Most notably right now, it will flag if the plugin is an Aircraft, Scenery, or Plugin.

I have also added an "example.XPMod.json" file to use as a very basic starting point.  Chances are this will stay this basic and not need to be expanded, but it could become more useful in the future.